test: file_processing function
